requirements.txt
datasets>=2.14
openai>=1.0
pandas>=2.0
scikit-learn>=1.3
numpy>=1.24
pyyaml>=6.0
tqdm>=4.65
pyarrow>=14.0
python-dotenv>=1.0
wandb>=0.16
scipy>=1.10
matplotlib>=3.8
dvc>=3.50
transformers>=4.36
torch>=2.1

# Testing
pytest>=8.0
pytest-cov>=5.0
pytest-mock>=3.12
